abstract out deprecateComponent
