+++ src/use-deprecated-actionmenu.js
@@ -1,31 +1,11 @@
-const { SyntaxKind } = require('ts-morph');
-const updateImportDeclaration = require('./utils/change-import-to-deprecated');
+const deprecateComponent = require('./utils/deprecate-component');
 
 const componentImportNames = ['ActionMenu', 'ActionMenuProps'];
 const fileName = 'ActionMenu';
 const v2FileNameToIgnore = 'ActionMenu2';
 
 const transform = (project) => {
-  const sourceFiles = project.getSourceFiles();
-
-  sourceFiles.forEach((sourceFile) => {
-    try {
-      sourceFile.getDescendantsOfKind(SyntaxKind.ImportDeclaration).forEach((declaration) => {
-        declaration = updateImportDeclaration(
-          declaration,
-          sourceFile,
-          componentImportNames,
-          fileName,
-          v2FileNameToIgnore
-        );
-      });
-
-      // save source back to file
-      sourceFile.saveSync();
-    } catch (error) {
-      console.log(error);
-    }
-  });
+  deprecateComponent(project, componentImportNames, fileName, v2FileNameToIgnore);
 };
 
 module.exports = transform;
