+++ src/use-deprecated-grid.js
@@ -1,24 +1,10 @@
-const { SyntaxKind } = require('ts-morph');
-const updateImportDeclaration = require('./utils/change-import-to-deprecated');
+const deprecateComponent = require('./utils/deprecate-component');
 
 const componentImportNames = ['Grid', 'GridProps'];
 const fileName = 'Grid';
 
 const transform = (project) => {
-  const sourceFiles = project.getSourceFiles();
-
-  sourceFiles.forEach((sourceFile) => {
-    try {
-      sourceFile.getDescendantsOfKind(SyntaxKind.ImportDeclaration).forEach((declaration) => {
-        declaration = updateImportDeclaration(declaration, sourceFile, componentImportNames, fileName);
-      });
-
-      // save source back to file
-      sourceFile.saveSync();
-    } catch (error) {
-      console.log(error);
-    }
-  });
+  deprecateComponent(project, componentImportNames, fileName);
 };
 
 module.exports = transform;
